# syncthreads before wmma